requirements.txt
Flask # O la versión más reciente
pymongo # O la versión más reciente compatible con tu Atlas
python-dotenv 
gunicorn